Refactor: extract infrastructure from server.ts and remove monolithic exports (fixes #49)
