+++ backend/server.ts
@@ -10,8 +10,8 @@ import cors from 'cors';
 import dotenv from 'dotenv';
 import path from 'path';
 import fs from 'fs';
-import { getDbPath, queryLocalDb, resetDbConnection } from './db.js';
-export { queryLocalDb };
+import { getDbPath, queryLocalDb, resetDbConnection } from './src/db/client.js';
+
 // @ts-ignore
 import api from '@actual-app/api';
 import AdmZip from 'adm-zip';
@@ -83,7 +83,7 @@ async function _doSyncInternal(): Promise<void> {
   }
 }
 
-export function doSync(): Promise<void> {
+function doSync(): Promise<void> {
   if (syncState.syncPromise) return syncState.syncPromise;
   syncState.syncPromise = _doSyncInternal().finally(() => { 
     syncState.syncPromise = null; 
@@ -97,76 +97,7 @@ function startPeriodicSync() {
   setInterval(doSync, 15 * 60 * 1000);
 }
 
-// Rotate backups: keep only the 10 most recent files
-function rotateBackups(backupDir: string) {
-  try {
-    const files = fs.readdirSync(backupDir);
-    const zipFiles = files
-      .filter(file => file.endsWith('.zip'))
-      .map(file => ({
-        name: file,
-        path: path.join(backupDir, file),
-        mtime: fs.statSync(path.join(backupDir, file)).mtime.getTime()
-      }));
-      
-    zipFiles.sort((a, b) => b.mtime - a.mtime);
-    
-    if (zipFiles.length > 10) {
-      const filesToDelete = zipFiles.slice(10);
-      for (const file of filesToDelete) {
-        fs.unlinkSync(file.path);
-        console.log(`[Backup] Deleted old backup file: ${file.name}`);
-      }
-    }
-  } catch (error) {
-    console.error('[Backup] Rotation error:', error);
-  }
-}
-
-// Perform budget backup
-export async function runBackup(): Promise<string> {
-  const dbPath = getDbPath();
-  const budgetDir = path.dirname(dbPath);
-  const backupDir = path.join(DATA_DIR, 'budget-backup');
-  
-  if (!fs.existsSync(backupDir)) {
-    fs.mkdirSync(backupDir, { recursive: true });
-  }
-  
-  const now = new Date();
-  const YYYY = now.getFullYear();
-  const MM = String(now.getMonth() + 1).padStart(2, '0');
-  const DD = String(now.getDate()).padStart(2, '0');
-  const HH = String(now.getHours()).padStart(2, '0');
-  const Min = String(now.getMinutes()).padStart(2, '0');
-  const SS = String(now.getSeconds()).padStart(2, '0');
-  
-  const suffix = Math.random().toString(36).substring(2, 6);
-  const filename = `${YYYY}${MM}${DD}_${HH}${Min}${SS}_${suffix}.zip`;
-  const zipPath = path.join(backupDir, filename);
-  
-  console.log(`[Backup] Starting backup to ${zipPath}...`);
-  
-  const zip = new AdmZip();
-  
-  const dbFile = path.join(budgetDir, 'db.sqlite');
-  if (fs.existsSync(dbFile)) {
-    zip.addLocalFile(dbFile);
-  } else {
-    throw new Error(`db.sqlite not found at ${dbFile}`);
-  }
-  
-  const metaFile = path.join(budgetDir, 'metadata.json');
-  if (fs.existsSync(metaFile)) {
-    zip.addLocalFile(metaFile);
-  }
-  
-  zip.writeZip(zipPath);
-  console.log(`[Backup] Backup written successfully: ${filename}`);
-  
-  rotateBackups(backupDir);
-  return filename;
-}
+import { runBackup } from './src/features/Backup/service.js';
 
 let lastBackupDateString = '';
 
@@ -219,53 +150,7 @@ app.post('/api/evaluate', (req, res) => {
   res.json({ result });
 });
 
-app.post('/api/backup', async (req, res) => {
-  try {
-    const filename = await runBackup();
-    res.json({ success: true, filename });
-  } catch (err: any) {
-    console.error("Manual backup failed:", err);
-    res.status(500).json({ error: err.message || String(err) });
-  }
-});
 
-app.get('/api/backups', (req, res) => {
-  try {
-    const backupDir = path.join(DATA_DIR, 'budget-backup');
-    if (!fs.existsSync(backupDir)) {
-      return res.json({ backups: [] });
-    }
-    const files = fs.readdirSync(backupDir);
-    const backups = files
-      .filter(file => file.endsWith('.zip'))
-      .map(file => {
-        const filePath = path.join(backupDir, file);
-        const stat = fs.statSync(filePath);
-        return {
-          filename: file,
-          size: stat.size,
-          createdAt: stat.mtime.toISOString()
-        };
-      });
-    backups.sort((a, b) => b.createdAt.localeCompare(a.createdAt));
-    res.json({ backups });
-  } catch (err: any) {
-    res.status(500).json({ error: err.message || String(err) });
-  }
-});
-
-app.get('/api/backups/download/:filename', (req, res) => {
-  const filename = req.params.filename;
-  const safeFilename = path.basename(filename);
-  const backupDir = path.join(DATA_DIR, 'budget-backup');
-  const filePath = path.join(backupDir, safeFilename);
-
-  if (fs.existsSync(filePath)) {
-    res.download(filePath);
-  } else {
-    res.status(404).json({ error: 'Backup file not found' });
-  }
-});
 
 // /api/data endpoint has been removed. Use vertical slice API endpoints (e.g. /api/dashboard) instead.
