+++ backend/src/features/BudgetEnvelope/service.ts
@@ -5,7 +5,7 @@
  */
 
 import type { IDbClient } from '../../infrastructure/db/IDbClient.js';
-import { defaultDbClient } from '../../../db.js';
+import { defaultDbClient } from '../../db/client.js';
 import type { EnvelopeHealth } from '../../../../shared/types/BudgetEnvelope.js';
 
 export class BudgetEnvelopeService {
